Add Listings to Homepage
The homepage was updated with listing data (address, city, price, desc, and up to 10 images) gathered from a Google Sheet API connection and shown in cards with a responsive image gallery. The background color of the homepage was darkened a bit to make the white listing cards pop. Text breaking was fixed with auto hyphens from github.com/vivgui/tailwindcss-hyphens.