Improve Color Consistency
The consistency of colors across the website was improved. Links now match the colors of buttons. The favicon color was also updated to match the main blue color. Also, the name/website-title area in both the page footer and on the homepage was updated to flex-row the image on desktop.